Quitada implementacion privada de UI
